requirements.txt
Django~=2.0.6
django-crispy-forms
pillow
djangorestframework
django-rest-swagger